requirements.txt
streamlit
numpy
tensorflow
Pillow
matplotlib
opencv-headless